refactor: consolidate version sync logic into single script
Merge sync-root-version.mjs functionality into version-with-root.mjs
to simplify the codebase. One script now handles both changesets
version update and root package sync.

- Remove separate sync-root-version.mjs file
- Include all logic in version-with-root.mjs
- Simpler and more maintainable
